Add event model and controller, add entry point
