+++ src/models/Event.php
@@ -1,0 +1,28 @@
+<?php
+/*  Riichi mahjong API game server
+ *  Copyright (C) 2016  o.klimenko aka ctizen
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 3 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+namespace Riichi;
+
+class EventModel extends Model
+{
+    public function getRatingTable($event, $orderBy)
+    {
+        if (!in_array($orderBy, ['name', 'rating', 'avg_place'])) {
+            throw new InvalidParametersException("Parameter orderBy should be either 'name', 'rating' or 'avg_place'");
+        }
+    }
+}
